requirements.txt
cryptography
SQLAlchemy
Click
PyInquirer
Faker
python-dotenv
